excise the codebase from some C99 format specifiers. (%zu -> %lu)
git-svn-id: https://warzone2100.svn.sourceforge.net/svnroot/warzone2100/trunk@7120 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/droid.c
@@ -3502,7 +3502,7 @@ UDWORD getDroidEffectiveLevel(DROID *psDroid)
 const char *getDroidNameForRank(UDWORD rank)
 {
 	ASSERT( rank < (sizeof(arrRank) / sizeof(struct rankMap)),
-	        "getDroidNameForRank: given rank number (%d) out of bounds, we only have %zu ranks\n", rank, (sizeof(arrRank) / sizeof(struct rankMap)) );
+	        "getDroidNameForRank: given rank number (%d) out of bounds, we only have %lu ranks\n", rank, (unsigned long) (sizeof(arrRank) / sizeof(struct rankMap)) );
 
 	return PE_("rank", arrRank[rank].name);
 }
